add filetype support for Elixir
autotag works great for me, but, it does not work in my Elixir files. However, it works perfectly for me when I include the elixir and heex filetypes here. In Elixir speak, the changes to this file allows me to have html tag closing enabled within "HEEx templates and .heex files". It does not enable html tag closing within ".eex" files which is fine as Elixir has moved to using .heex files instead of the older .eex files.
Much Respect!