Add eruby defaults (#104)
* Fix minimal test setup

Inclusion of any config with `-u` also adds `~/.config/nvim/pack/*/start/*`
to rtp, which is probably the reason for tests passing on developers'
machines. This "problem" remains, and can in fact break tests through
interference with other plugins.

* Add eruby to default list of filetypes

This merely adds tests and one line of data to default setup values.